+++ HISTORY.rst
@@ -17,6 +17,7 @@ History
 * Uniformize the use of matrix k_ and add an argument "ensemble" to method "predict" in regression.py
 * Add replication of the Chen Xu's tutorial testing Jackknife+aB vs Jackknife+
 * Add Jackknife+-after-Bootstrap documentation
+* Add EnbPI method for Time Series Regression
 * Improve scikit-learn pipelines compatibility
 
 0.3.1 (2021-11-19)